fix: move .env settings to .ts
