--- .env.example
@@ -1,4 +1,0 @@
-HOST=http://localhost
-FRONTEND_PORT=3000
-BACKEND_PORT=3001
-ENV=dev
