+++ src/server/index.ts
@@ -1,22 +1,20 @@
 import express from 'express';
-import { config } from 'dotenv';
-import { CORS, DLL_PATH, ROUTES } from './settings';
+import { CORS, ROUTES } from './config';
 import { OperationResponse } from './types/operation-response.type';
 import { OperationRequest } from './types/operation-request.type';
-
-config();
+import { BACKEND_PORT, HOST } from '../settings';
 
 const app = express();
 app.use(CORS);
 
 for (const route in ROUTES) {
-  app.get(route, (req: OperationRequest, res: OperationResponse): void => {
-    const { first, second } = req.query;
-    let result = String(ROUTES[route]([+first, +second]));
-    res.json(result);
-  });
+	app.get(route, (req: OperationRequest, res: OperationResponse): void => {
+		const { first, second } = req.query;
+		let result = String(ROUTES[route]([+first, +second]));
+		res.json(result);
+	});
 }
 
-app.listen(process.env.BACKEND_PORT, () => {
-  console.log(`Backend started: ${process.env.HOST}:${process.env.BACKEND_PORT}`);
+app.listen(BACKEND_PORT, () => {
+	console.log(`Backend started: ${HOST}:${BACKEND_PORT}`);
 });
