+++ src/server/load-DLL.ts
@@ -1,0 +1,31 @@
+import { DataType, define, open } from 'ffi-rs';
+import { DLL_PATH } from './config';
+
+export const loadDLL = (name: string, path: string) => {
+	open({
+		library: name,
+		path: DLL_PATH,
+	});
+	return define({
+		NM_addition: {
+			library: name,
+			retType: DataType.Double,
+			paramsType: [DataType.Double, DataType.Double],
+		},
+		NM_subtraction: {
+			library: name,
+			retType: DataType.Double,
+			paramsType: [DataType.Double, DataType.Double],
+		},
+		NM_multiplication: {
+			library: name,
+			retType: DataType.Double,
+			paramsType: [DataType.Double, DataType.Double],
+		},
+		NM_division: {
+			library: name,
+			retType: DataType.Double,
+			paramsType: [DataType.Double, DataType.Double],
+		},
+	});
+};
