+++ src/settings.ts
@@ -1,0 +1,3 @@
+export const HOST = 'http://localhost';
+export const FRONTEND_PORT = 3000;
+export const BACKEND_PORT = 3001;
